feat!: Overhaul CSS classes and allow specifying custom key types
Make classes of drawn rect's and text's better composable. Remove literal
restrictions from `type` field and add the same field to combo specs.
This would allow users to specify a custom type and define it in the CSS
if they'd like custom styling.

Also fixes multi-line alignment logic to use `dy` property with `em` unit spacing.

This is a breaking change in the config in the config since some classes were renamed
and `line_spacing` meaning changed to actually correspond to line spacing.

Co-authored-by: Manna Harbour <manna-harbour@users.noreply.github.com>