feat: Auto-shrink legends that are too wide
Instead of text overflowing to the sides, make it so that
the font size is scaled down for wide legends. It is enabled
by default but can be disabled via config.

This is currently not an ideal design, since the
default value is only tuned for the default config settings and
it doesn't translate very well to combo boxes. Ideally
it would consider the font size and rect/combo box width
and tune itself automatically but the current CSS styling
approach doesn't allow getting the font size easily.