feat: Consistently assign "key"/"combo" CSS classes
Make sure to assign "key" to layer key SVG elts and "combo" to combo elts.
We no longer need to use the `type` "combo" for combos, so remove that default.

We will ideally switch to removing classes from the rect/text elts and keep them
only in the <g> container, and this requires changing the default CSS to use
descendant selectors.