fix: Combine shrink_wide_legends with CSS font size
Previously, if shrink_wide_legends took effect for a legend,
it would completely override the font size rather than scaling
the underlying font size (usually specified in CSS). Use a child
element with the style attr so that they properly combine.

This was an issue for too long hold/shifted legends where their
font could get larger rather than smaller when the config kicked in.

Thanks to @inxi in Glove80 Discord for pointing out the problem and
proposing the solution.