feat: Implement QMK's ISO return, hackily
When a key is exactly 1.25u x 2.0u, QMK converts it to a ISO
return key which has an extra 0.25u protrusion to the left on its
top half. Implement that via an internal flag that triggers the
draw of an extra rect with a special mask so it blends together
with the original key rect. It probably doesn't work great for some
CSS overrides and with draw_config.draw_key_sides.